requirements.txt
# pip install -r requirements.txt
flask
psycopg2
bcrypt
gunicorn
python-dotenv
requests